# Try to fix "enforce return conventions" perf regression