Reenabled some rewrite cases, though that seemed to not really matter.  Looks like most of the regression came from section ordering.

```
           django_template3.py             2.6s (2)             2.6s (2)  -0.3%
                 pyxl_bench.py             2.2s (2)             2.1s (2)  -2.0%
     sqlalchemy_imperative2.py             2.5s (2)             2.5s (2)  -0.2%
       django_template3_10x.py            16.0s (2)            15.8s (2)  -1.1%
             pyxl_bench_10x.py            16.9s (2)            16.6s (2)  -1.5%
 sqlalchemy_imperative2_10x.py            19.4s (2)            19.5s (2)  +0.5%
                       geomean                 6.5s                 6.4s  -0.8%
```